Moving Balance interface to interface folder
