+++ src/app/interfaces/common/Balance.ts
@@ -1,0 +1,8 @@
+export interface Balance {
+  coin: string;
+  free: number;
+  spotBorrow: number;
+  total: number;
+  usdValue: number;
+  availableWithoutBorrow: number;
+}
