+++ src/app/store/features/depositSlice.ts
@@ -1,5 +1,6 @@
 import { createSlice } from '@reduxjs/toolkit';
-import { Balance, getBalanceFulfiled } from 'infrastructure/services/deposit/DepositService';
+import { Balance } from 'app/interfaces/common/Balance';
+import { getBalanceFulfiled } from 'infrastructure/services/deposit/DepositService';
 
 const initialState: DepositState = {
   currentBalance: 0,
